更新日志 v2.4（大概）
性能优化
延迟图片验证：资产扫描时不再完整加载图片验证，仅检查文件存在性
虚拟滚动列表：干员列表改用固定 8 槽位动态更新，减少 80% 内存占用
Bug 修复
存储显示溢出：修复大于 4GB 存储显示负数问题（uint32→uint64）
Arc 进度条：修复显示可用空间而非已用空间的问题
改进
自动单位换算：存储容量自动显示 MB/GB/TB 单位

+++ src/ui/actions_sysinfo.h
@@ -9,9 +9,9 @@ void ui_sysinfo_get_meminfo_str(char *ret, size_t ret_sz);
 void ui_sysinfo_get_os_release_str(char *ret, size_t ret_sz);
 uint32_t ui_sysinfo_get_file_crc32(const char *path);
 bool ui_sysinfo_is_sdcard_inserted();
-uint32_t ui_sysinfo_get_nand_available_size();
-uint32_t ui_sysinfo_get_sd_available_size();
-uint32_t ui_sysinfo_get_nand_total_size();
-uint32_t ui_sysinfo_get_sd_total_size();
+uint64_t ui_sysinfo_get_nand_available_size();
+uint64_t ui_sysinfo_get_sd_available_size();
+uint64_t ui_sysinfo_get_nand_total_size();
+uint64_t ui_sysinfo_get_sd_total_size();
 int ui_sysinfo_format_sd_card();
 // EEZ回调不需要添加。
